[#20871]CDCSDK: Fix issue of empty batch being sent in case of ALTER TABLE failure
Summary:
In the case when alter table add column fails, we receive a `CHANGE_METADATA_OP` with incorrect schema. In the current implementation, the method populate CDCSDKIntentRecord() relies on the comparison of column count with the column count in schema for populating a new proto record. Since the incorrect schema has more columns then the intents received, no records are populated and we end up sending an empty batch of records.

This diff changes the decision making logic for populating a new proto record. Instead of relying on the column count, we now check the next intent and compare its primary key and hybrid time with the current intent's primary key and hybrid time.
Jira: DB-9859

Test Plan:
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestAtomicDDLRollback
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestUpdateIntentsWithIncompleteRows
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestInsertIntentsWithIncompleteRows
All other CDC tests

Reviewers: skumar, asrinivasan, stiwary, devansh.saxena

Reviewed By: skumar

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D32103