[#21127] YSQL: Use new framework to pass --ysql_enable_db_catalog_version_mode
Summary:
Currently the gflag --ysql_enable_db_catalog_version_mode is passed from tserver
to PG backend via environment variable. During testing sometimes we need to
change the default value of this gflag but that is not very convenient. It
should be as simple as changing the default value (from false to true) at the
place where this gflag is defined:

```
DEFINE_NON_RUNTIME_PREVIEW_bool(ysql_enable_db_catalog_version_mode, false,
    "Enable the per database catalog version mode, a DDL statement that only "
    "affects the current database will only increment catalog version for "
    "the current database.");
```

However in reality, we also must change all the places such as:

```
    if (YBCIsEnvVarTrueWithDefault(
            "FLAGS_ysql_enable_db_catalog_version_mode", false) &&
```
If any such place is missed, it can result in test failure that takes debugging
effort (which actually happened).

This diff changes this gflag to use the new framework of passing NON_RUNTIME
gflag from tserver to PG backend so the usage in PG of this gflag will be
`*YBCGetGFlags()->ysql_enable_db_catalog_version_mode`.

Test Plan:
./yb_build.sh --cxx-test pg_catalog_version-test

Did manual micro benchmark testing showed no meaningful change in  performance
of the function YBIsDBCatalogVersionMode(), in particular during the tryout
phase in cluster upgrade to a release where this gflag is turned on by default.
During the tryout phase the gflag is on but the table `pg_yb_catalog_version`
isn't migrated to have one row per database yet.

              instr_time total = {0};
              instr_time start;
              instr_time end;
              INSTR_TIME_SET_CURRENT(start);
              for (int i = 0; i < 100000; ++i)
                      YBIsDBCatalogVersionMode();
              INSTR_TIME_SET_CURRENT(end);
              INSTR_TIME_ACCUM_DIFF(total, end, start);
              int64 microsecs = total.tv_sec * 1e6 + total.tv_nsec / 1e3;
              YBC_LOG_INFO("call_time=%.3fus", (double)microsecs);

Reviewers: jason

Reviewed By: jason

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D32567