[PLAT-12321] Feature flag for release decoupling
Summary:
Added a new feature flag to allow enabling the release decoupling
work. Also wrote a wrapper around ReleaseMetadata and Release to allow
ReleaseManager and its clients to be able to easily work with
both implementations.

Test Plan:
with the feature flag disabled:
1. create universe (vm and k8s)
2. upgrade univers (vm and k8s)
3. add new releases
4. delete releases

no testing with flag enabled, it is not yet supported

Reviewers: muthu, anijhawan

Reviewed By: muthu

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32653