[#18228] YSQL: mitigate backends catalog version RPC timeouts
Summary:
In some setups, the master-tserver WaitForYsqlBackendsCatalogVersion RPC
can sometimes continually timeout.  This usually happens if the postgres
connection startup takes too long, usually because master is located far
away and catalog caching flags are not optimally set for such a
geo-distributed cluster.  The main flag is
ysql_enable_read_request_caching, which will be set true by default
soon.  In that case, this issue may mostly disappear, but still tie up
any loose ends.

- Improve the error message.  Before, it is like

      ERROR:  timed out waiting for postgres backends to catch up
      DETAIL:  -1 backends on database 16385 are still behind catalog version 43.
      HINT:  Run the following query on all tservers to find the lagging backends: SELECT * FROM pg_stat_activity WHERE catalog_version < 43 AND datid = 16385;

  Now, make it like

      ...
      DETAIL:  Failed to determine how many backends on database 16385 are still behind catalog version 43.

  in case of num_lagging_backends == -1 (which signifies master didn't
  successfully RPC at least one tserver).
- Add a master gflag
  wait_for_ysql_backends_catalog_version_master_tserver_rpc_timeout_ms
  to control the master-tserver WaitForYsqlBackendsCatalogVersion RPC
  timeout specifically.  Before, it is controlled by the generic
  master_ts_rpc_timeout_ms flag, which affects many other RPCs.  With a
  separate flag, if the user hits the timeout, they can increase this
  flag to possibly avoid it without affecting other RPCs.  Also, set the
  default for this timeout to 60s rather than the previous 30s, which
  should address the case when ysql_enable_read_request_caching is false
  and the cluster is geo-distributed.  Update some tests due to the
  timeout change.

In the future, another improvement could be to "cache" one PG connection
on each tserver for use by these WaitForYsqlBackendsCatalogVersion RPCs
so that they do not have to deal with connection startup catalog
caching.  There could be a small queue to access this one connection.
It should generally be faster than the current approach if, after the
initial catalog caching, the remaining waiters in the queue can run
their quick local query, but it is possible to take longer than the
current approach in case there are frequent catalog version bumps
happening, in which case each waiter would have to catalog cache
refresh on top of waiting in a queue.
Jira: DB-7238

Test Plan:
On Almalinux 8:

    ./yb_build.sh fastdebug --gcc11 \
      --cxx-test pgwrapper_pg_backends-test

Close: #18228
Backport-through: 2.21.0

Reviewers: myang

Reviewed By: myang

Subscribers: ybase, yql, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32601