[#21181] yugabyted-ui: Improve the master and tserver address list caching in the API layer
Summary:
Currently, there are many functions in the yugabyted-ui api server that require a list of master/tserver addresses.
For example, a function may need to query a master/tserver, but it doesn't care which one.
It uses a list of addresses so that it can attempt a different address if the first request fails.

This diff implements a local cache of master/tserver addresses to improve performance.
The caches are implemented as slices of strings protected by a read-write lock.
The master cache updates when we get a successful response from :7000/api/v1/masters or :9000/api/v1/masters.
It also updates when all http requests made to a list of cached master addresses fails.
The tserver cache updates when we get a successful response from :7000/api/v1/tablet-servers.
It also updates when all attempts to get a ysql connection using a list of cached tserver addresses fails.

Included in this diff are also minor bugfixes and improvements.
A more detailed list of changes:
- Removed unused `nodeHost` parameter from `GetMastersFuture`
- Improved tserver address caching to cache a slice of addresses instead of just one
- Tserver address cache will automatically update whenever `GetTabletServersFuture` returns successfully
- Added master address caching similar to tserver address caching that caches a slice of addresses
- Master address cache will update automatically whenever `GetMastersFuture` or `GetMastersFromTserverFuture` return successfully
- Added function `BuildMasterURLsAndAttemptGetRequests` that first uses cached master addresses, and calls `AttemptGetRequests`. If this fails, calls `BuildMasterURLs` to fetch new master addresses, and calls `AttemptGetRequests` again.
- Refactored code that sequentially called `BuildMasterURLs` and `AttemptGetRequests` to call `BuildMasterURLsAndAttemptGetRequests` instead
- Change `BuildMasterURLs` to accept url query parameters as an argument
- Fix `AttemptGetRequests` to properly handle situations where `expectJson` is true and response status is OK but response is not valid json
- Added mutex locks to protect creation of gocql sessions/pgxpool connections, for the case where two api calls both try to make a new connection at the same time

Test Plan: no test plan

Reviewers: nikhil

Reviewed By: nikhil

Subscribers: yugabyted-dev, djiang

Differential Revision: https://phorge.dev.yugabyte.com/D32625