[#21135] docdb: Fix AddStartingTablet to only check missing replicas after LB init
Summary:
`AddStartingTablet` used to mark a newly started tablet as over-replicated if the tablet is not missing replicas (i.e. is not found in `tablets_missing_replicas_`). This was being called during `UpdateTabletInfo` (before the missing replicas map was initialized, which is incorrect) and during regular load balancing when adding a replica (which is correct).

This was causing tablets with starting replicas to be marked as over-replicated and have replicas removed even if the tablet was not over-replicated.

This diff adds an `initialized_` field to the load balancer `PerTableLoadState` which is set after `UpdateTabletInfo` is called and changes `AddStartingTablet` to only perform the over-replicated check if `initialized_` is set. It also `SCHECK`s `initialized_` in some other functions that we expect to only be called before / after initialization. This necessitated some test clean up.
Jira: DB-10075

Test Plan: `ybd --cxx-test master_load_balancer_mocked-test --gtest_filter LoadBalancerMockedTest.TestStartingTablet` and existing LB tests

Reviewers: jhe

Reviewed By: jhe

Subscribers: slingam, ybase, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32629