[#21161] YSQL: Return snapshot_name in response to CREATE_REPLICATION_SLOT command
Summary:
This is related to the project to support Replication slot API in YSQL (#18724).
(https://phorge.dev.yugabyte.com/D29194).
This is also related to the PG Compatible Logical Replication Consumption project.

In response to the CREATE_REPLICATION_SLOT command, the server will send a one-row result set.
One of the fields in this result set is "snapshot_name". In PG, this refers to the identifier of the
snapshot exported by the command. The snapshot consumption then happens through the execution of a
SELECT query as of the consistent state referred to by "snapshot_name". In YB terms, it would be the
equivalent of querying the database with a specified read point.

The YB support for the CREATE_REPLICATION_SLOT command returns the uint64 representation of the
HybridTime corresponding to the consistent snapshot time of the consistent snapshot stream that
was created as part of the CREATE_REPLICATION_SLOT command. This support is added in this revision.

**UPGRADE/ROLLBACK SAFETY:**
These changes are protected via the preview flag: ysql_yb_enable_replication_commands

The following are the PB changes. These are all for RPC responses.

1 yb::master::CreateCDCStreamResponsePB - addition of 1 optional field
2 yb::tserver::PgCreateReplicationSlotResponsePB - addition of 1 optional field
Jira: DB-10095

Test Plan:
```
./yb_build.sh --cxx-test cdcsdk_consistent_snapshot-test
./yb_build.sh --cxx-test cdcsdk_snapshot-test
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressReplicationSlot'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot'
```

```
Added new test
./yb_build.sh --cxx-test integration-tests_cdcsdk_consistent_snapshot-test --gtest_filter CDCSDKConsistentSnapshotTest.TestSnapshotNameFromCreateReplicationSlot
```

Reviewers: skumar, stiwary, xCluster, hsunder, aagrawal

Reviewed By: stiwary

Subscribers: ycdcxcluster, ybase, yql, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32610