[PLAT-12824] Use the current AMI ID for getting the devices information if passed
Summary:
AWS does not updates the ImageId in the metadata post root volume replacement, returning old ImageID.
This leads to subsequent failure in case the old AMI is deleted from the cloud console. This diff passes the AMI
ID from the java layer & instead uses the same to describe the devices in case passed with fallback to metadata
ImageId.

Test Plan:
Created an AMI on AWS.
Deployed universe with the same.
Deleted the AMI from the AWS console.
Perfomed OS Patching using the new AMI (Verified that the node is running the updated OS)

Reviewers: muthu, yshchetinin, amalyshev

Reviewed By: muthu

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32616