[BACKPORT 2.20][PLAT-12478][PLAT-8586][PLAT-10924][PLAT-11868][PLAT-11810][PLAT-11783][PLAT-10547][PLAT-12608] AZ overridable Proxy settings in Universe, YBC client-server version upgrade, Backports to support Proxy changes
Summary:
[BACKPORT 2.20][PLAT-10547] Editing universe to same no of nodes is allowed via /universe_configure and PUT /universe/uuid

Right now it is possible to submit edit page without any changes, and it takes some time to complete that task.
And also it is possible to submit edit with changes, that will not be applied correctly (for example we can add one node and simultaneously
change volume size - this will lead to inconsistent state as userIntent will be updated but only new node will have updated volume)
Current diff adds multiple validations to edit flow:
1) Validate that changes could be applied through EditUniverse (for example not allowing volume increase for VMs unless it is full move)
2) Validate that some there are no side-changes (like updating gflags)

[BACKPORT 2.20][PLAT-11783] [2.21.0.0-b312] [testcloudgeopartition] Add Universe Region failed [Error: No changes that could be applied by EditUniverse]

Previously we added a precheck to edit universe that doesn't allow NoOp.
But this is too trict - thus we restrict adding region without actual modifying placement.
We should allow that because this operation was used previously.

[BACKPORT 2.20][PLAT-11810] Can't edit the RF on RR

We should consider changing RF as a possible option for edit universe

[BACKPORT 2.20][PLAT-11868]: Can't edit the number of cores/volume on the K8 universe

If we try changing the number of cores/volume size of master/tserver on the K8 universe and observe that it fails and throws ‘No changes could be
applied with the edit universe’
Hence we need to check if tservernode spec or masternodespec has changed

[BACKPORT 2.20][PLAT-10924] When creating the universe object, we should have `updateSucceeded` bit default to false

Setting updateSucceeded to false before universe creation task to avoid green mark in UI in case
when subsequent create task fails at the validation phase

[BACKPORT 2.20][PLAT-8586] Add validation to EditUniverse

Added validation that nodePrefix is unchanged during edit universe

[BACKPORT 2.20][PLAT-12478]AZ overridable Proxy settings in Universe

- Design doc: https://docs.google.com/document/d/1XRHfhrLPlteOzjDrkgRtbc9lQf8FYSwjBu4l6QJyy4Q/edit
- Added AZ overridable proxy settings in Universe user Intent
- Modified YBC based backup/restore code to use proxy settings configured at universe, falling back to S3 storage config settings for backward compatibility
- The API request with zone overrides looks like the following:
```
POST    /customers/:cUUID/universes/:uniUUID/upgrade/proxy_config

"proxyConfig": {
	"httpProxy": "10.9.131.7:3128",
    "httpsProxy": "10.9.131.7:3128",
    "noProxyList": ["a.b.c"]
}
"userIntentOverrides": {
  "azOverrides": {
    "4a720488-d71b-40f7-9816-9f0566f58803": {
    	"instanceType": "c5.large",
    	"proxyConfig": {
      		"httpProxy": "10.9.131.7:3128",
      		"httpsProxy": "10.9.131.7:3128"
      	}
    },
    "c10ee796-f28b-4ec5-a226-14a4b6ab1cfa": {
    	"instanceType": "c5.large",
    	"proxyConfig": {
      		"httpProxy": "10.9.131.7:3128",
      		"httpsProxy": "10.9.131.7:3128"
      	}
    }
  }
}
```
- The API modifies Proxy Config default and AZ overrides and does not affect any other settings in the User Intent.
- The set of default proxy and az overrides provided in the API request is saved in the User Intent.
- Added validation checks on Proxy:
  # `httpProxy` and `httpsProxy` both cannot be null at the same time.
  # `httpProxy` and `httpsProxy` should have correct url format: `<scheme>://<username>:<password>@<ip>:<port>`.
  # `scheme` defaults to `http`
  # `port` is optional defaulting to 0
  # `username` and `password` default to null values.

[BACKPORT 2.20][PLAT-12608] Upgrade ybc version to 2.1.0.0-b4: Adds Proxy Support for Azure, Support for proxy and no-proxy, Build YBC with glibc 2.17

Increment the YBC client and server version to 2.1.0.0-b4
- https://github.com/yugabyte/ybc/commit/ba31116940b2f6dcbde1474ec811c720c253da16 - Add support for proxy for connections to Azure.
- https://github.com/yugabyte/ybc/commit/877346eeec96c6a9260108f91b154aa6edf44eb0 - Add support for no-proxy when specifying a proxy for AWS and Azure
- https://github.com/yugabyte/ybc/commit/e9def1a1bcf0232c5e4e3a2f2db9ae5c690f68fb - Build YBC with lower glibc version

ISSUE [PLAT-12520]

YBC currently does not have support for specifying a proxy to talk to Azure Storage.

FIX

Added the support for specifying a proxy for Azure. Also added the plumbing to get the proxy info all the way to the
google code but not implemented the proxy for google for now. Technically we support both http and https for talking
to the proxy but we will only be testing http due to lack of time. Added support in the cli to pick up the proxy url form
the env variable YBC_HTTP_PROXY. For example:

export YBC_HTTP_PROXY=http://vkumar:vkumar@172.151.20.153:3128

TESTING

Setup a squid http proxy and tested the code by specifying the proxy in a universe. Ran some of the UTs that does backup to S3, Azure and google storage after specifying the YBC_HTTP_PROXY env variable and verified that those pass.

ISSUE [PLAT-12520]

We need to be able to support no_proxy so that on a given node can download from one cloud through the proxy while for the other cloud it can do the same without a proxy. This is because, during restore, each node will need to get data from more than one cloud in a multi region configuration where some tablets are backed up to one cloud while others to another cloud or to another region on the same cloud).

In addition, there is currently no way to send the proxy or any other certificate to YBC.

FIX
Add a ProxyConfig in the CloudStoreConfig that has one default proxy spec and a per node proxy spec where you can define the proxy host and a list of hosts that don't need a proxy.
In addition add a additional_certs filed to CloudStoreConfig that the user can use to send the proxy and other certificates needed.
The yb-controller-cli now takes a credentials_file that contains the json representation of the CloudStoreConfig to enable cli users to easily specify multi region configs including proxy and additional certificates settings.

TESTING

Tested by running against a squid proxy. Tested with both a http enabled proxy and another https enabled proxy.

Json format of the credentials_file

{
  "defaultSpec": {
    "type": "S3",
    "creds": {
      "AWS_ACCESS_KEY_ID": "ACCESS KEY",
      "AWS_DEFAULT_REGION": "us-west-2",
      "AWS_SECRET_ACCESS_KEY": "SECRET KEY"
    },
    "bucket": "csunittest"
  },
  "regionSpecMap": {
    "us-east1": {
      "type": "AZURE",
      "creds": {
        "AZURE_STORAGE_END_POINT": "https://kkgtest.blob.core.windows.net",
        "AZURE_STORAGE_SAS_TOKEN": "SAS TOKEN"
      },
      "bucket": "csunittest"
    }
  },
  "proxyConfig": {
    "nodeProxyMap": {
      "bchan-db1": {
      },
      "bchan-db2": {
      },
      "bchan-db3": {
        "scheme": "http",
        "host": "100.127.128.95",
        "port": 3128
      }
    }
  },
  "additionalCerts": "<add in the certificates here>"
}

ISSUE [PLAT-12598]

Make sure that ybc is built with a compatible GLIBC version based on the YBDB builds after move to non-Linux brew builds.

TESTING

Unit tests pass.

Test Plan: Dev itests

Reviewers: #yba-api-review!, cwang, sanketh, rmadhavan, yshchetinin, anijhawan, nsingh, vpatibandla, sneelakantan, kkg

Reviewed By: yshchetinin, kkg

Subscribers: yugaware, sneelakantan

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D32701