[PLAT-12608] Upgrade ybc version to 2.1.0.0-b4: Adds Proxy Support for Azure, Support for proxy and no-proxy, Build YBC with glibc 2.17
Summary:
Increment the YBC client and server version to 2.1.0.0-b4
- https://github.com/yugabyte/ybc/commit/ba31116940b2f6dcbde1474ec811c720c253da16 - Add support for proxy for connections to Azure.
- https://github.com/yugabyte/ybc/commit/877346eeec96c6a9260108f91b154aa6edf44eb0 - Add support for no-proxy when specifying a proxy for AWS and Azure
- https://github.com/yugabyte/ybc/commit/e9def1a1bcf0232c5e4e3a2f2db9ae5c690f68fb - Build YBC with lower glibc version

ISSUE [PLAT-12520]

YBC currently does not have support for specifying a proxy to talk to Azure Storage.

FIX

Added the support for specifying a proxy for Azure. Also added the plumbing to get the proxy info all the way to the
google code but not implemented the proxy for google for now. Technically we support both http and https for talking
to the proxy but we will only be testing http due to lack of time. Added support in the cli to pick up the proxy url form
the env variable YBC_HTTP_PROXY. For example:

export YBC_HTTP_PROXY=http://vkumar:vkumar@172.151.20.153:3128

TESTING

Setup a squid http proxy and tested the code by specifying the proxy in a universe. Ran some of the UTs that does backup to S3, Azure and google storage after specifying the YBC_HTTP_PROXY env variable and verified that those pass.

ISSUE [PLAT-12520]

We need to be able to support no_proxy so that on a given node can download from one cloud through the proxy while for the other cloud it can do the same without a proxy. This is because, during restore, each node will need to get data from more than one cloud in a multi region configuration where some tablets are backed up to one cloud while others to another cloud or to another region on the same cloud).

In addition, there is currently no way to send the proxy or any other certificate to YBC.

FIX
Add a ProxyConfig in the CloudStoreConfig that has one default proxy spec and a per node proxy spec where you can define the proxy host and a list of hosts that don't need a proxy.
In addition add a additional_certs filed to CloudStoreConfig that the user can use to send the proxy and other certificates needed.
The yb-controller-cli now takes a credentials_file that contains the json representation of the CloudStoreConfig to enable cli users to easily specify multi region configs including proxy and additional certificates settings.

TESTING

Tested by running against a squid proxy. Tested with both a http enabled proxy and another https enabled proxy.

Json format of the credentials_file

{
  "defaultSpec": {
    "type": "S3",
    "creds": {
      "AWS_ACCESS_KEY_ID": "ACCESS KEY",
      "AWS_DEFAULT_REGION": "us-west-2",
      "AWS_SECRET_ACCESS_KEY": "SECRET KEY"
    },
    "bucket": "csunittest"
  },
  "regionSpecMap": {
    "us-east1": {
      "type": "AZURE",
      "creds": {
        "AZURE_STORAGE_END_POINT": "https://kkgtest.blob.core.windows.net",
        "AZURE_STORAGE_SAS_TOKEN": "SAS TOKEN"
      },
      "bucket": "csunittest"
    }
  },
  "proxyConfig": {
    "nodeProxyMap": {
      "bchan-db1": {
      },
      "bchan-db2": {
      },
      "bchan-db3": {
        "scheme": "http",
        "host": "100.127.128.95",
        "port": 3128
      }
    }
  },
  "additionalCerts": "<add in the certificates here>"
}

ISSUE [PLAT-12598]

Make sure that ybc is built with a compatible GLIBC version based on the YBDB builds after move to non-Linux brew builds.

TESTING

Unit tests pass.

Test Plan: dev itests

Reviewers: agunta, asharma, kkg

Reviewed By: kkg

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32594