[PLAT-12843] Implement addJsonPathToLeafNodes utility function
Summary:
Adds a new utility function which transforms given JsonNode. It
replaces the leaf node's value with an object containing jsonPath to
that node and its value. This will primarily be used by Kubernetes
Provider validation but might be useful to other provider validation
implementations as well.

This basically helps to have full JSONPath of the field being
validated in the error message sent to the front-end/user.

Test Plan:
Tested the method with full provider object and it works as
expected. This change should not have impact on anything else as it is
unused as of now.

Reviewers: anijhawan, dshubin, rohita.payideti

Reviewed By: dshubin, rohita.payideti

Subscribers: asharma, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32636