[#20920] YSQL, ASH: Update yb_active_session_history column names
Summary:
This diff updates two column names for the view yb_active_session_history.

`yql_endpoint_tserver_uuid` is updated to `top_level_node_id`, based on the
feedback that the former was not intuitive.

`session_id` is updated to `ysql_session_id` as this is ysql field.

The migration script which introduced the view is updated as there have been
no releases with this script.

Jira: DB-9902

Test Plan:
./yb_build.sh --java-test TestYbAsh
./yb_build.sh --java-test TestYsqlUpgrade
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressRules#testPgRegressRules'

Reviewers: jason

Reviewed By: jason

Subscribers: hbhanawat, amitanand, yql

Differential Revision: https://phorge.dev.yugabyte.com/D32646