[PLAT-12855] Updating Almalinux 8 images
Summary:
Issue with GPG key on older Almalinux 8 default images we had. This diff updates the generate script
to be a little easier to use (and gets rid of Azure ARM support) and reruns it with 8.9 as the
desired version.

Test Plan:
create universe on aws/gcp/azure, ensure create universe succeeds and new packages
(glibc-langpack-en) can be downloaded on the OS.

Reviewers: svarshney, prarabdh.garg

Reviewed By: prarabdh.garg

Subscribers: nbhatia, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32652