[#21217] YSQL: revert epoch code in YsqlBackendsManager
Summary:
A series of commits touch YsqlBackendsManager since the initial
foundation:

1. d7cf0a74421676eb7271089bc186aabe0fdb7c85
1. f8580549cb07767241d84355a5524bc417b16722
1. 10dfd6cfa95e4fa61c02bcf27a92b57e71a9c802

The reason for the first commit is invalidated by the last commit, so
clean up any leftover code and revert the whole thing.  This means
getting rid of unused epoch parameter in LaunchTS and changing epoch
back to term because epoch is no longer needed.
Jira: DB-10145

Test Plan:
On Almalinux 8:

    ./yb_build.sh fastdebug --gcc11 --cxx-test pgwrapper_pg_backends-test

Close: #21217

Reviewers: asrivastava

Reviewed By: asrivastava

Subscribers: zdrudi, ybase, bogdan, hsunder

Differential Revision: https://phorge.dev.yugabyte.com/D32663