[#21217] YSQL: revert epoch code in YsqlBackendsManager
Summary:
A series of commits touch YsqlBackendsManager since the initial
foundation:

1. 3a7537a9d6d2a4c7a1fc7fae5fca997ee4d35569
1. 85e5c2e391bdff3b3495890821fcddaad313addf
1. 0753a759544fc65527967d19080240701efdf84b

The reason for the first commit is invalidated by the last commit, so
clean up any leftover code and revert the whole thing.  This means
getting rid of unused epoch parameter in LaunchTS and changing epoch
back to term because epoch is no longer needed.
Jira: DB-10145

Test Plan:
On Almalinux 8:

    ./yb_build.sh fastdebug --gcc11 --cxx-test pgwrapper_pg_backends-test

Close: #21217

Reviewers: asrivastava

Reviewed By: asrivastava

Subscribers: zdrudi, ybase, bogdan, hsunder

Differential Revision: https://phorge.dev.yugabyte.com/D32663