[#10650] YSQL: tweak index backfill timeout flags
Summary:
Change index backfill timeout-related flags to reduce the chance of
hitting timeout-related failures, particularly for YSQL.

- master --ysql_index_backfill_rpc_timeout_ms: 1m -> 5m
- tserver --backfill_index_timeout_grace_margin_ms:
  - when --backfill_index_rate_rows_per_sec is 0 (or less) (default):
    - YSQL: 1s -> 3m
    - YCQL: 1s
  - otherwise:
    - YSQL: 2 * batch_time + 1s -> max(2 * batch_time, 3m)
    - YCQL: 2 * batch_time + 1s -> max(2 * batch_time, 1s)

Before, timeout failures typically occurred due to spending more time in
a batch than the grace margin.  A grace margin of 1 second is extremely
low to account for the all the back and forth communication in YSQL:

- master
  - tserver
    - postgres (timeout check here, so worst case we have grace margin
      amount of time left to respond)
      - tserver (for read)
    - postgres
      - tserver (for write)
    - postgres
  - tserver
- master (we should have gotten here within grace margin amount of time)

Conditions that worsen the situation include geo-distributed clusters
and GIN indexes, which can take a long time during write because a batch
of X could expand to N * X writes, where N is how many elements are
typically in the indexed column.

Some local testing with a GIN index of N=10000 rf=1 cluster revealed it
could take as long as two minutes to complete a batch.  For that reason,
increase the margin to three minutes (for YSQL).  Alternatively, the
batch size default could have been fixed to be lowered to 128 (see issue
#10649), but testing showed a noticeable performance degradation in that
case.

Consequently, the overall YSQL timeout needs to be increased to be
higher than that margin, so set it to five minutes.  This means, that
the master UI tasks and the pg_stat_progress_create_index view refresh
less frequently: at least every two minutes (for each ongoing backfill
chunk) compared to approximately one minute before.
Jira: DB-1126

Test Plan:
On Almalinux 8:

    #!/usr/bin/env bash
    set -eux

    # Initial build.
    ./yb_build.sh release daemons initdb

    # Reverting flags causes test to fail.
    set +e
    output=$(YB_EXTRA_MASTER_FLAGS='--ysql_index_backfill_rpc_timeout_ms=60000' \
      YB_EXTRA_TSERVER_FLAGS='--backfill_index_timeout_grace_margin_ms=1000' \
      ./yb_build.sh release --gtest_filter PgIndexBackfillTest.GinStress)
    ret=$?
    set -e
    [ "$ret" -ne 0 ]
    grep "Reached maximum number of retries (0) for request Backfilling indexes { iii }" \
      <<<"$output"

    # Test passes 1000 times.
    ./yb_build.sh release --gtest_filter PgIndexBackfillTest.GinStress \
      -n 1000 --tp 1

Although the test new test PgIndexBackfillTest.GinStress works in
fastdebug, it fails for debug, and there is no existing macro to enable
only for fastdebug and release, so go the lazy way of forgetting about
fastdebug and only enabling the test for release.

Close: #10650

Reviewers: amitanand

Reviewed By: amitanand

Subscribers: ybase, yql, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32631