[#21159] DocDB: Include indexed_table_id along with an index
Summary:
The change includes the indexed_table_id along with the index when listing tables using ListTablesRequestPB.

This way to determine the main table associated with an index, a second lookup is not required.

**Upgrade/Rollback safety:**
The change adds a new optional field to the ListTablesResponsePB. There are no consumers of this field currently other than the test, so we expect this to be upgrade-safe.
Jira: DB-10093

Test Plan: ybt master_master-test MasterTest.TestListTablesIncludesIndexedTableId

Reviewers: jhe, sanketh, xCluster, hsunder

Reviewed By: jhe, sanketh

Subscribers: ybase, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32609