[#21187] YSQL: Support SET yb_read_time TO '<hybridTime_uint64> ht'
Summary:
Currently, the value to which the yb_read_time variable is to be set is to be a Unix timestamp in microseconds.
(https://phorge.dev.yugabyte.com/D26192)
This revision allows for specifying the value alternatively as the uint64 representation of a HybridTime.

This motivation for this enhancement comes from the PG Compatible Logical Replication Consumption project,
specifcally the Snapshot Consumption part.

With this revision, it will be possible to specify the value in two ways. A unit ht will be used to distinguish
between the two cases. This is similar to the approach used to support a Numeric with Unit type variable.
The yb_read_time variable will continue to be a String variable.

Usage Examples -

```
SET yb_read_time TO 1686095506000000
```

This is the existing support and will be the default. The value here is UNIX timestamp in microseconds

```
SET yb_read_time TO '6990255045012508672 ht'
```

The value here is the uint64 representation of HybridTime. The ht unit is used to identify this.

As with Numeric with Unit type variables, the value must be written as a string (with quotes) to use this feature.
The unit name is case-sensitive, and there can be whitespace between the numeric value and the unit.
(https://www.postgresql.org/docs/11/config-setting.html)
Jira: DB-10120

Test Plan:
```
Enhanced exsting test
ybd --cxx-test pgwrapper_pg_read_time-test --gtest_filter PgMiniTestBase.CheckReadingDataAsOfPastTime
```

Reviewers: mhaddad, pjain, stiwary, skumar, aagrawal

Reviewed By: mhaddad, pjain

Subscribers: slingam, bogdan, ycdcxcluster, ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D32634