[#21190] CDCSDK: Add cdc_state entry for replication slot
Summary:
This diff is part of the PG compatible logical replications. It introduces a new entry in cdc_state table for a replication slot as well as some additional fields in a cdc_state row to hold extra metadata for the slot.

This primary requirement of this entry is persist some slot-related metadata that will be used by the Virtual WAL for initialisation of LSN & txnID generators. VirtualWAL will also update this entry when it receives feedback from the walsender.

The cdc_state entry for replication slot will be added during stream creation only if the following conditions are met:
1. Stream is created with a consistent_snapshot stream option (USE_SNAPSHOT/NOEXPORT_SNAPSHOT)
2. FLAGS_ysql_yb_enable_replication_commands is set to true.

New fields introduced are all present under the **data** column:
1. Confirmed_flush_lsn (uint64)
2. Restart_lsn (uint64)
3. RecordId's commit_time (uint64) [This represent the CDCSDKUniqueRecordID's commit_time]
4. Xmin (uint32)

Similar to existing fields of the data column, these new fields will be stored as optional & serialized to string in the cdc_state table.

The cdc_state entry for the replication slot will be initialised to the following values in each column:
1. Tablet_id: set to "dummy_id_for_replication_slot"
2. stream_id: as per actual stream_id associated with the slot
3. Checkpoint: not set
4. Data:

  - active_time: not set
  - cdc_sdk_safe_time: set to consistent snapshot_time
  - snapshot_key: not set
  - confirmed_flush_lsn: 2 (representing consistent point)
  - restart_lsn: 1
  - xmin: 1
  - record_id_commit_time: set to consistent snapshot_time

5. Last replication time: not set

**Note: **In UpdatePeersAndMetrics, we ignore this entry for replication slot while deciding to move retention barriers.

Testing
New tests (TestCDCStateEntryForReplicationSlot) has been added to verify presence of slot entry in cdc_state only in case the 2 conditions (mentioned above) are met.
Jira: DB-10123

Test Plan: ./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestCDCStateEntryForReplicationSlot

Reviewers: stiwary, asrinivasan, skumar

Reviewed By: stiwary, asrinivasan

Subscribers: ybase, ycdcxcluster, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32643