[PLAT-12082]: Add mechanism to replace node
Summary:
The goal is to allow YBA to remove unreachable/unhealthy nodes from the universe, specially for onprem providers and to introduce simplified node ops compared to the current workflow.

1. New convenience actions on the universe nodes page.
  a. Replace Node = Remove + Release + Add
  c. In summary, universe nodes will show the following actions
       i. Stop Processes
      ii. Start Processes
     iii. Replace Node
      iv. Decommission Node
       v. Advanced section will have Remove/Release/Add/Delete Node actions. None of the current actions are being removed.

Test Plan:
Please refer to the video
{F150095}

NOTE: @cwang is working on backend diff, so waiting for merging D31991

NOTE: @cwang is working on adding Node state for DECOMMISSION NODE in the response of OnPremNode list. I have a UI diff, but will wait for Charles to create a diff in that case

Reviewers: jmak, cwang

Reviewed By: jmak

Subscribers: cwang, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32726