[#21249] YSQL: lower catalog lease period
Summary:
The YSQL Backends Manager uses the concept of a "lease" to bound the
amount of time it waits on unresponsive tservers before giving up and
deeming them to be resolved.  This "lease" was initially conveniently
implemented using existing flag tserver_unresponsive_timeout_ms (default
one minute), which is tied to TSDescriptor::IsLive.  Although YSQL
Backends Manager bounds the amount of time it waits for an unresponsive
tserver by this much, the tserver never respected this "lease" as it was
left as a todo (see issue #13369).

An issue was found where new master leaders ought to always wait out the
lease period before serving YSQL Backends Manager requests.  In that
case, a one minute lease is a very long time to wait for.  To reduce the
waiting period, create a new flag master_ts_ysql_catalog_lease_ms to
represent the "lease", change all code to refer to this new flag, and
reduce this number to ten seconds.  Make the flag a preview flag since
it is still up for change/discussion.  Ten seconds was chosen because it
is small enough that it won't make too much impact during leader changes
and high enough that leases won't expire due to network flakiness or
slowdown (at least I hope).  The lease is still refreshed using
heartbeats.  There are still open issues that will be solved in the
future:

- tserver does not respect the lease (issue #13369)
- master does not wait out the lease period after being newly elected as
  leader (issue #21251)

Also remove some test classes reducing the lease period to 30s or 10s
because the new default is now 10s.
Jira: DB-10170

Test Plan:
On Almalinux 8:

    ./yb_build.sh fastdebug --gcc11 --cxx-test pgwrapper_pg_backends-test

Close: #21249
Depends on D32771

Reviewers: myang

Reviewed By: myang

Subscribers: ybase, yql, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32737