[PLAT-12822] Set correct sshUser/Port for node-agent context
Summary:
As part of OS Patching, when we move the base AMI for example, from alma -> ubuntu the base SSH user changes.
we persist the imageBundle information in universe at the end of task therefore userIntent still points to the old bundle thus the old user, which fails during NodeAgent installation.

This diff sets the sshUser/Port in the ShellContext for OSPatching so that we use the same

Test Plan: Performed OS Patching from alma to ubuntu

Reviewers: nsingh

Reviewed By: nsingh

Subscribers: Shubham, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32614