[BACKPORT 2.20][#21251] YSQL: wait out lease period in backends manager
Summary:
When a master becomes newly elected as leader (including cluster
startup), the new master initially is unaware of any tservers, so
GetAllDescriptors returns empty.  Then, the BackendsCatalogVersionJob
will never make any progress because it normally makes progress through
the HandleResponse of a response from a master-tserver RPC.  This was
hit about 5% of the time in TSAN testing with tests such as
PgBackendsTest.CacheLost because TSAN is slow enough to sometimes get in
this situation where, upon launching a job on the new master, no
tservers have registered yet.

This further reveals a theoretical issue with network partitioned
tservers.  In case there is a tserver that becomes network partitioned
coincidently with a master leader change, the new master leader won't
know about this tserver.  Therefore, we need master to wait out the
lease period before being able to accurately respond to any
WaitForYsqlBackendsCatalogVersion requests. The main consequence of this
is a default ten seconds lag on CREATE INDEX statements whenever the
master leader changes or on cluster startup.  Some tests using indexes
slow down as well.

Implement this by sending back Uninitialized status in the client-master
RPC.  Make YBClient interpret this as a case of retrying with
exponential backoff.  This requires exposing ShouldRetry to
SyncClientMasterRpc, specializing it only for the
WaitForYsqlBackendsCatalogVersion RPC.

Several tests, particularly PgBackendsTest.*, need some adjustment
because of the new behavior.  Mainly solve this by adding a helper
function WaitOutInitialCatalogLeasePeriod and using it in the tests that
need it.
Jira: DB-10172
Original commit: e2dc9812c85d04839c2b7aa7c1c4929e204394bd / D32739

Test Plan:
This command fails before this fix:

    ./yb_build.sh tsan --clang17 \
      --cxx-test pgwrapper_pg_backends-test \
      --gtest_filter 'PgBackendsTest.CacheLost' \
      -n 50

Check for PgBackendsTest.* regressions using release build:

    DEFAULT_TEST_TIMEOUT_SEC=99999 ./yb_build.sh release --clang17 \
      --cxx-test pgwrapper_pg_backends-test \
      -n 50

Check for other test regressions on Detective.

Depends on D39628

Reviewers: myang, mlillibridge

Reviewed By: mlillibridge

Subscribers: asrivastava, mlillibridge

Differential Revision: https://phorge.dev.yugabyte.com/D39629