[#21134] xCluster: Fix race condition with XClusterPoller shutdown and StoreReplicationError
Summary:
On xClusterPoller shutdown, we can hit a scenario where we have triggered the shutdown and removed the poller from our error_map, but then are still handling parts of the callback, where we will try to add any errors to the error_map.
This would then hit a DCHECK that was checking that the poller exists in the error_map. Removing this DCHECK and just handling this case.
Jira: DB-10074

Test Plan:
```
ybd --cxx-test xcluster_ddl_replication-test --gtest_filter "XClusterDDLReplicationTest.CreateTable" -n 100
```

Reviewers: hsunder, xCluster

Reviewed By: hsunder

Subscribers: ybase, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32575