[#21267] YSQL: Fix failing test TestPgParallelReadIsolation
Summary:
The test did not account for the fact that the Read Committed isolation
is disabled by default in release builds and since the test requires
Read Committed properties it has to enable it by Gflag.

Also there is a timing issue. Due to release build being faster, it is
highly probable that the read and the write transaction are to close
and read transaction fails upon receiving the restart error. Added
artificial 1 second delay to prevent this.
Jira: DB-10185

Test Plan: ybd --java-test org.yb.pgsql.TestPgParallelReadIsolation

Reviewers: pjain

Reviewed By: pjain

Subscribers: bogdan, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D32834